added basic Template for Load Testing
