+++ src/main/java/com/revature/templates/LoadTestConfig.java
@@ -1,0 +1,7 @@
+package com.revature.templates;
+
+public class LoadTestConfig {
+    public int loops;
+    public int threads;
+    public int rampUp;
+}
