+++ src/test/java/com/revature/docutest/TestUtil.java
@@ -7,6 +7,10 @@ public class TestUtil {
     
     public static Swagger swag1;
     
+    {
+        initFields();
+    }
+    
     public static void initFields() {  
         swag1 = new SwaggerParser().read("src/test/resources/example.json");
     }
